added animals resources counter and HP(stamina)
today I worked hard on dawn realm, i added 4 different animals, cow, sheep, chicken and pig. to them I gave the same run away mechanism as the slime. I also added an HP bar on each and every character and it is displayed when player enters the detection range and when the HP bar get to zero % the character dies and is removed from screen. I also added some resources counters to check how much inventory you got, added some wood, gold and silver which will be used in the store and building tomorrow i will work on either player death and the story or on the music and SFX